All commands are refactored based on session feedback
