+++ src/Coreutils-EYAD_HUSSEIN/cmd/tail/main.go
@@ -1,0 +1,14 @@
+package main
+
+import (
+	"Coreutils-EYAD_HUSSEIN/commands"
+	"Coreutils-EYAD_HUSSEIN/models"
+	"Coreutils-EYAD_HUSSEIN/utils"
+)
+
+func main() {
+	flags := utils.GetFlags()
+	utils.ValidateFlags(flags, models.CommandsMap["tail"].Flags)
+
+	commands.Tail(flags)
+}
