+++ src/Coreutils-EYAD_HUSSEIN/commands/tree.go
@@ -29,13 +29,13 @@ func Tree(flags []string) {
 		maxDepth = tempMaxDepth
 	}
 
-	for index, arg := range os.Args[2:] {
+	for index, arg := range os.Args[1:] {
 		if arg == "-L" || arg == strconv.Itoa(maxDepth) {
 			continue
 		}
 		printDirectory(arg, 0, maxDepth)
 
-		if index != len(os.Args[2:])-1 {
+		if index != len(os.Args[1:])-1 {
 			fmt.Println()
 		}
 	}
